SOLR-1024 -- Calling abort on DataImportHandler import commits data instead of calling rollback
git-svn-id: https://svn.apache.org/repos/asf/lucene/solr/trunk@745831 13f79535-47bb-0310-9956-ffa450edef68

+++ contrib/dataimporthandler/src/main/java/org/apache/solr/handler/dataimport/DataImporter.java
@@ -462,7 +462,7 @@ public class DataImporter {
 
     public int start = 0;
 
-    public int rows = Integer.MAX_VALUE;
+    public long rows = Integer.MAX_VALUE;
 
     public boolean clean = true;
